Fix: Player
1. Player Entity 내에는 실행 메서드를 작성하지 않고, 이를 Service 로 모아서 작성
2. Service 에는 숫자 입력받기, 종료 숫자 입력받기 작성
